process data installation + images

+++ FromSofwareStorage/DatabaseEngineProvider.cs
@@ -1,6 +1,9 @@
 ﻿using System;
 using System.Data.SqlServerCe;
 using System.IO;
+using System.Text;
+using System.Threading.Tasks;
+using FromSoftwareStorage.Images;
 
 namespace FromSoftwareStorage
 {
@@ -11,6 +14,9 @@ namespace FromSoftwareStorage
         private const string BackupDatabaseName = "Data_backup.sdf";
         private const string RsaPrivateKeyFile = "DataPrivateKey.key";
         private const string RsaPublicKeyFile = "DataPublicKey.key";
+
+        private readonly ASCIIEncoding _bytesConverter = new ASCIIEncoding();
+        private readonly string[] _sqlSeparator = { "\r\nGO\r\n", "\r\nGO", "\r\nGo\r\n", "\r\nGo", "\r\ngo\r\n", "\r\ngo", "\r\ngO\r\n", "\r\ngO" };
 
         public DatabaseEngineProvider()
         {
@@ -31,19 +37,20 @@ namespace FromSoftwareStorage
 
         public bool HasPublicKey => File.Exists(RsaPrivateKeyFileName);
 
-        public void CreateDatabase(string password)
+        public void CreateDatabase(string password = null)
         {
             GenerateNewKeys(password);
             CreateDatabaseWithName(DatabaseFileName, password);
         }
 
-        public void CreateDatabase()
-        {
-            CreateDatabaseWithName(DatabaseFileName);
-        }
-
         public void GenerateNewKeys(string password)
         {
+            if (string.IsNullOrEmpty(password))
+                return;
+
+            if (string.IsNullOrWhiteSpace(password))
+                return;
+
             if(HasPrivateKey)
                 File.Delete(RsaPrivateKeyFileName);
 
@@ -51,6 +58,94 @@ namespace FromSoftwareStorage
                 File.Delete(RsaPublicKeyFileName);
 
             EncryptPassword(password);
+        }
+
+        public async Task<bool> InstallDatabaseAsync(string password = null)
+        {
+            string connectionString = BuildConnectionString(DatabaseFileName, password);
+            if (!DataBaseExists)
+            {
+                GenerateNewKeys(password);
+                DoCreation(connectionString);
+            }
+
+            if (await ExecuteSql(connectionString))
+            {
+                return await InsertImagesAsync(connectionString);
+            }
+
+            return false;
+        }
+
+        private async Task<bool> ExecuteSql(string connectionString)
+        {
+            using (SqlCeConnection sqlCeConnection = new SqlCeConnection(connectionString))
+            {
+                sqlCeConnection.Open();
+
+                using (SqlCeTransaction sqlCeTransaction = sqlCeConnection.BeginTransaction())
+                {
+                    byte[] sqlBytes = await SqlEngineProvider.GetSqlDataBytesAsync();
+                    string sqlData = _bytesConverter.GetString(sqlBytes);
+                    string[] sqlStatements = sqlData.Split(_sqlSeparator, StringSplitOptions.RemoveEmptyEntries);
+
+                    foreach (var sqlStatement in sqlStatements)
+                    {
+                        using (SqlCeCommand sqlCeCommand = sqlCeConnection.CreateCommand())
+                        {
+                            sqlCeCommand.Transaction = sqlCeTransaction;
+                            sqlCeCommand.CommandText = sqlStatement;
+                            sqlCeCommand.ExecuteNonQuery();
+                        }
+                    }
+
+                    sqlCeTransaction.Commit();
+                }
+
+                return true;
+            }
+        }
+
+        private async Task<bool> InsertImagesAsync(string connectionString)
+        {
+            string sqlCommand = await SqlEngineProvider.GetImageInsertCommandAsync();
+
+            using (SqlCeConnection sqlCeConnection = new SqlCeConnection(connectionString))
+            {
+                sqlCeConnection.Open();
+
+                using (SqlCeTransaction sqlCeTransaction = sqlCeConnection.BeginTransaction())
+                {
+                    // dark souls3
+                    byte[] darkSouls3Resource = await ImageResource.GetDarkSouls3ImageResourceAsync();
+                    ExecuteSqlCommandWithParameters("DarkSouls 3", darkSouls3Resource, sqlCeConnection, sqlCommand, sqlCeTransaction);
+
+                    // Sekiro
+                    byte[] sekiroResource = await ImageResource.GetSekiroImageResourceAsync();
+                    ExecuteSqlCommandWithParameters("Sekiro", sekiroResource, sqlCeConnection, sqlCommand, sqlCeTransaction);
+                    
+                    sqlCeTransaction.Commit();
+
+                    return true;
+                }
+            }
+        }
+
+        private static void ExecuteSqlCommandWithParameters(string gameName,
+            byte[] imageFile, 
+            SqlCeConnection sqlCeConnection, 
+            string sqlCommand,
+            SqlCeTransaction sqlCeTransaction)
+        {
+            using (SqlCeCommand sqlCeCommand = sqlCeConnection.CreateCommand())
+            {
+                sqlCeCommand.CommandText = sqlCommand;
+                sqlCeCommand.Parameters.AddWithValue("@GameName", gameName);
+                sqlCeCommand.Parameters.AddWithValue("@ImageFile", imageFile);
+
+                sqlCeCommand.Transaction = sqlCeTransaction;
+                sqlCeCommand.ExecuteNonQuery();
+            }
         }
 
         private void EncryptPassword(string password)
@@ -85,6 +180,5 @@ namespace FromSoftwareStorage
         {
             DoCreation(BuildConnectionString(databaseName, password));
         }
-
     }
 }
